feat: 특정 시간에 대한 예약이 존재하는데, 그 시간을 삭제하려 할 때 예외가 발생하도록 구현

+++ src/main/java/roomescape/reservationTime/service/ReservationTimeService.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import roomescape.globalException.CustomException;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservationTime.ReservationTimeMapper;
 import roomescape.reservationTime.domain.ReservationTime;
 import roomescape.reservationTime.domain.dto.ReservationTimeReqDto;
@@ -15,10 +16,12 @@ import roomescape.reservationTime.repository.ReservationTimeRepository;
 public class ReservationTimeService {
 
     private final ReservationTimeRepository repository;
+    private final ReservationRepository reservationRepository;
 
     @Autowired
-    public ReservationTimeService(ReservationTimeRepository repository) {
+    public ReservationTimeService(ReservationTimeRepository repository, ReservationRepository reservationRepository) {
         this.repository = repository;
+        this.reservationRepository = reservationRepository;
     }
 
     public List<ReservationTimeResDto> readAll() {
@@ -36,7 +39,10 @@ public class ReservationTimeService {
     }
 
     public void delete(Long id) {
-        repository.findByIdOrThrow(id);
+        ReservationTime reservationTime = repository.findByIdOrThrow(id);
+        if (reservationRepository.existsByReservationTime(reservationTime)) {
+            throw new CustomException(HttpStatus.BAD_REQUEST, "예약에서 사용 중인 시간입니다.");
+        }
         repository.delete(id);
     }
